Made publish validation happen always
On the theory that this validation is meant for development as an ongoing check for runtime errors, I made validation happen even if no one is listening to the given channel. This way you don't have to set up a listener to test that the event you're emitting is valid.